Fixed return types to uint8

+++ contracts/api.sol
@@ -5,11 +5,11 @@ contract DateTimeAPI {
          */
         function isLeapYear(uint16 year) constant returns (bool);
         function getYear(uint timestamp) constant returns (uint16);
-        function getMonth(uint timestamp) constant returns (uint16);
-        function getDay(uint timestamp) constant returns (uint16);
-        function getHour(uint timestamp) constant returns (uint16);
-        function getMinute(uint timestamp) constant returns (uint16);
-        function getSecond(uint timestamp) constant returns (uint16);
+        function getMonth(uint timestamp) constant returns (uint8);
+        function getDay(uint timestamp) constant returns (uint8);
+        function getHour(uint timestamp) constant returns (uint8);
+        function getMinute(uint timestamp) constant returns (uint8);
+        function getSecond(uint timestamp) constant returns (uint8);
         function getWeekday(uint timestamp) constant returns (uint8);
         function toTimestamp(uint16 year, uint8 month, uint8 day) constant returns (uint timestamp);
         function toTimestamp(uint16 year, uint8 month, uint8 day, uint8 hour) constant returns (uint timestamp);
